Moved sample_rupts in event_based
